+++ Source/SourceExpander.Embedder/AssemblyUtil.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using System.Reflection;
 
-namespace SourceExpander
+namespace SourceExpander;
+
+internal class AssemblyUtil
 {
-    internal class AssemblyUtil
-    {
-        public static readonly Version AssemblyVersion = Assembly.GetExecutingAssembly().GetName().Version;
-    }
+    public static readonly Version AssemblyVersion = Assembly.GetExecutingAssembly().GetName().Version;
 }